+++ apps/controllerx/cx_core/controller.py
@@ -586,5 +586,21 @@ class Controller(Hass, Mqtt):  # type: ignore[misc]
         """
         return None
 
+    def get_shelly_actions_mapping(self) -> Optional[DefaultActionsMapping]:
+        """
+        Controllers can implement this function. It should return a dict
+        with the command that a controller can take and the functions as values.
+        This is used for Shelly support.
+        """
+        return None
+
+    def get_shellyforhass_actions_mapping(self) -> Optional[DefaultActionsMapping]:
+        """
+        Controllers can implement this function. It should return a dict
+        with the command that a controller can take and the functions as values.
+        This is used for Shelly for HASS support.
+        """
+        return None
+
     def get_predefined_actions_mapping(self) -> PredefinedActionsMapping:
         return {}